+++ gate/base/mqtt_agent.go
@@ -70,7 +70,7 @@ func (this *agent) OnInit(gate gate.Gate, conn network.Conn) error {
 	this.isclose = false
 	this.rev_num = 0
 	this.send_num = 0
-	this.last_storage_heartbeat_data_time=time.Duration(time.Now().UnixNano())
+	this.last_storage_heartbeat_data_time = time.Duration(time.Now().UnixNano())
 	return nil
 }
 func (a *agent) IsClosed() bool {
@@ -212,7 +212,7 @@ func (a *agent) recoverworker(pack *mqtt.Pack) {
 		if r := recover(); r != nil {
 			buff := make([]byte, 4096)
 			runtime.Stack(buff, false)
-			log.Error("Gate recoverworker error [%v] stack : %v",r, string(buff))
+			log.Error("Gate recoverworker error [%v] stack : %v", r, string(buff))
 		}
 	}()
 
@@ -315,34 +315,34 @@ func (a *agent) recoverworker(pack *mqtt.Pack) {
 			}
 		}
 		//if a.GetSession().GetUserId() != "" {
-			//这个链接已经绑定Userid
-			a.lock.Lock()
-			interval := int64(a.last_storage_heartbeat_data_time) + int64(a.gate.Options().Heartbeat) //单位纳秒
-			a.lock.Unlock()
-			if interval < time.Now().UnixNano() {
-				if a.gate.GetStorageHandler() != nil {
-					a.lock.Lock()
-					a.last_storage_heartbeat_data_time = time.Duration(time.Now().UnixNano())
-					a.lock.Unlock()
-					a.gate.GetStorageHandler().Heartbeat(a.GetSession())
-				}
+		//这个链接已经绑定Userid
+		a.lock.Lock()
+		interval := int64(a.last_storage_heartbeat_data_time) + int64(a.gate.Options().Heartbeat) //单位纳秒
+		a.lock.Unlock()
+		if interval < time.Now().UnixNano() {
+			if a.gate.GetStorageHandler() != nil {
+				a.lock.Lock()
+				a.last_storage_heartbeat_data_time = time.Duration(time.Now().UnixNano())
+				a.lock.Unlock()
+				a.gate.GetStorageHandler().Heartbeat(a.GetSession())
 			}
+		}
 		//}
 	case mqtt.PINGREQ:
 		//客户端发送的心跳包
 		//if a.GetSession().GetUserId() != "" {
-			//这个链接已经绑定Userid
-			a.lock.Lock()
-			interval := int64(a.last_storage_heartbeat_data_time) + int64(a.gate.Options().Heartbeat) //单位纳秒
-			a.lock.Unlock()
-			if interval < time.Now().UnixNano() {
-				if a.gate.GetStorageHandler() != nil {
-					a.lock.Lock()
-					a.last_storage_heartbeat_data_time = time.Duration(time.Now().UnixNano())
-					a.lock.Unlock()
-					a.gate.GetStorageHandler().Heartbeat(a.GetSession())
-				}
+		//这个链接已经绑定Userid
+		a.lock.Lock()
+		interval := int64(a.last_storage_heartbeat_data_time) + int64(a.gate.Options().Heartbeat) //单位纳秒
+		a.lock.Unlock()
+		if interval < time.Now().UnixNano() {
+			if a.gate.GetStorageHandler() != nil {
+				a.lock.Lock()
+				a.last_storage_heartbeat_data_time = time.Duration(time.Now().UnixNano())
+				a.lock.Unlock()
+				a.gate.GetStorageHandler().Heartbeat(a.GetSession())
 			}
+		}
 		//}
 	}
 }
