--- network/ws_client.go
@@ -1,145 +1,0 @@
-// Copyright 2014 mqant Author. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-package network
-
-import (
-	"github.com/gorilla/websocket"
-	"github.com/liangdas/mqant/log"
-	"sync"
-	"time"
-)
-
-type WSClient struct {
-	sync.Mutex
-	Addr             string
-	ConnNum          int
-	ConnectInterval  time.Duration
-	PendingWriteNum  int
-	MaxMsgLen        uint32
-	HandshakeTimeout time.Duration
-	AutoReconnect    bool
-	NewAgent         func(*WSConn) Agent
-	dialer           websocket.Dialer
-	conns            WebsocketConnSet
-	wg               sync.WaitGroup
-	closeFlag        bool
-}
-
-func (client *WSClient) Start() {
-	client.init()
-
-	for i := 0; i < client.ConnNum; i++ {
-		client.wg.Add(1)
-		go client.connect()
-	}
-}
-
-func (client *WSClient) init() {
-	client.Lock()
-	defer client.Unlock()
-
-	if client.ConnNum <= 0 {
-		client.ConnNum = 1
-		log.Warning("invalid ConnNum, reset to %v", client.ConnNum)
-	}
-	if client.ConnectInterval <= 0 {
-		client.ConnectInterval = 3 * time.Second
-		log.Warning("invalid ConnectInterval, reset to %v", client.ConnectInterval)
-	}
-	if client.PendingWriteNum <= 0 {
-		client.PendingWriteNum = 100
-		log.Warning("invalid PendingWriteNum, reset to %v", client.PendingWriteNum)
-	}
-	if client.MaxMsgLen <= 0 {
-		client.MaxMsgLen = 4096
-		log.Warning("invalid MaxMsgLen, reset to %v", client.MaxMsgLen)
-	}
-	if client.HandshakeTimeout <= 0 {
-		client.HandshakeTimeout = 10 * time.Second
-		log.Warning("invalid HandshakeTimeout, reset to %v", client.HandshakeTimeout)
-	}
-	if client.NewAgent == nil {
-		log.Error("NewAgent must not be nil")
-	}
-	if client.conns != nil {
-		log.Warning("client is running")
-	}
-
-	client.conns = make(WebsocketConnSet)
-	client.closeFlag = false
-	client.dialer = websocket.Dialer{
-		HandshakeTimeout: client.HandshakeTimeout,
-	}
-}
-
-func (client *WSClient) dial() *websocket.Conn {
-	for {
-		conn, _, err := client.dialer.Dial(client.Addr, nil)
-		if err == nil || client.closeFlag {
-			return conn
-		}
-
-		log.Warning("connect to %v error: %v", client.Addr, err)
-		time.Sleep(client.ConnectInterval)
-		continue
-	}
-}
-
-func (client *WSClient) connect() {
-	defer client.wg.Done()
-
-reconnect:
-	conn := client.dial()
-	if conn == nil {
-		return
-	}
-	conn.SetReadLimit(int64(client.MaxMsgLen))
-
-	client.Lock()
-	if client.closeFlag {
-		client.Unlock()
-		conn.Close()
-		return
-	}
-	client.conns[conn] = struct{}{}
-	client.Unlock()
-
-	wsConn := newWSConn(conn)
-	agent := client.NewAgent(wsConn)
-	agent.Run()
-
-	// cleanup
-	wsConn.Close()
-	client.Lock()
-	delete(client.conns, conn)
-	client.Unlock()
-	agent.OnClose()
-
-	if client.AutoReconnect {
-		time.Sleep(client.ConnectInterval)
-		goto reconnect
-	}
-}
-
-func (client *WSClient) Close() {
-	client.Lock()
-	client.closeFlag = true
-	for conn := range client.conns {
-		conn.Close()
-	}
-	client.conns = nil
-	client.Unlock()
-
-	client.wg.Wait()
-}
